+++ src/main/java/domain/Attendance.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.policy.AttendanceState;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
