+++ src/main/java/domain/policy/TimePolicy.java
@@ -1,4 +1,4 @@
-package policy;
+package domain.policy;
 
 import java.time.LocalTime;
