+++ src/test/java/config/AttendanceSheetFactoryTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
-import policy.AbsentPolicy;
+import domain.policy.AbsentPolicy;
 
 import java.io.IOException;
 import java.nio.file.Files;
