+++ src/test/java/domain/policy/ExpellStateTest.java
@@ -1,6 +1,6 @@
-package policy;
+package domain.policy;
 
-import org.junit.jupiter.api.BeforeEach;
+import domain.policy.ExpellState;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -9,7 +9,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static policy.ExpellState.*;
+import static domain.policy.ExpellState.*;
 
 public class ExpellStateTest {
