+++ src/test/java/domain/policy/TimePolicyTest.java
@@ -1,5 +1,6 @@
-package policy;
+package domain.policy;
 
+import domain.policy.TimePolicy;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
